// Front/src/app/model/Contract.ts
//import { Payment } from "../model/payment";

export class Contract {
     IDContract: string; 
    CreationDate: Date;
    StartDate: Date;
    EndDate: Date;
    NetPremium: number;
    TTCPremium: number;
    CeillingAmount: number;
    InstallmentsNB: number;
    Costatus: string;
    Method: string;
    IDClientP: string;
    IDClientC: string;
    //Payments : Payment
    
    }